+++ src/main/java/net/brnbrd/delightful/compat/Mods.java
@@ -13,6 +13,7 @@ import java.util.function.Supplier;
 public class Mods {
 	public static final String AA = "additionaladditions";
 	public static final String AE = "aether";
+	public static final String AE2 = "ae2";
 	public static final String AN = "ars_nouveau";
 	public static final String BOP = "biomesoplenty";
 	public static final String BTA = "botania";