+++ src/main/java/net/brnbrd/delightful/data/gen/DelightfulBlockStateProvider.java
@@ -30,6 +30,7 @@ public class DelightfulBlockStateProvider extends BlockStateProvider {
     protected void registerStatesAndModels() {
         this.wildCropBlock(DelightfulBlocks.WILD_SALMONBERRIES.get());
         this.stageBlock(DelightfulBlocks.SALMONBERRY_BUSH.get(), SalmonberryBushBlock.AGE);
+        this.stageBlock(DelightfulBlocks.CANTALOUPE_PLANT.get(), CantaloupePlantBlock.AGE);
         this.sackBlock(DelightfulBlocks.SALMONBERRY_SACK.get(), "salmonberry");
         this.sackBlock(DelightfulBlocks.ACORN_SACK.get(), "acorn");
         this.miniBlock((MiniBlock) DelightfulBlocks.CANTALOUPE.get(),