+++ src/main/java/net/brnbrd/delightful/data/gen/DelightfulRecipeProvider.java
@@ -512,7 +512,7 @@ public class DelightfulRecipeProvider extends RecipeProvider implements IConditi
             .addResultWithChance(Items.SUGAR, 0.5F, 1)
             .build(finished, Util.rl(Delightful.MODID, "cutting/sugar_cane"));
         CuttingBoardRecipeBuilder.cuttingRecipe(Ingredient.of(DelightfulItems.WILD_SALMONBERRIES.get()), Ingredient.of(ForgeTags.TOOLS_KNIVES), DelightfulItems.SALMONBERRY_PIPS.get(), 1)
-            .addResultWithChance(Items.ORANGE_DYE, 0.5F, 2)
+            .addResult(Items.ORANGE_DYE, 1)
             .build(finished, Util.rl(Delightful.MODID, "cutting/wild_salmonberries"));
         }